添加了五个全新的特定环境的对话

+++ chat/utils.py
@@ -22,7 +22,7 @@ def load_whisper_model():  # 实现模型的预加载
         "load_whisper_model method is called, model is loading...This model is 2.9G in size and will take 3-5 minutes to download for the first time access.")
     global WHISPER_MODEL
     model_size = "large-v2"
-    WHISPER_MODEL = WhisperModel(model_size, device="cuda", compute_type="float16")  # float16
+    WHISPER_MODEL = WhisperModel(model_size, device="cuda", compute_type="float32")  # float16
     print("Model successfully loaded!")
